backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
Pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
django-filter==23.5
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1